get resource pool without name
This appears to be the best way to allow cloning into a chosen cluster which is not using any resource pools. Luckily the cluster is able to return a default resource pool. I have this working with vsphere 5.1 and rbvmomi 1.6.0.
The option to the vm_clone method would look like this 'resource_pool' => ['cluster5', nil]